requirements.txt
cerebras_cloud_sdk
parallel-web
python-dotenv
requests
beautifulsoup4
streamlit
